Delete CNAME
I'm trying things out here so this might not work out. However, I want to remove it so that sub-projects, such as seeker, are not redirected to the main domain, which current doesn't work. By doing this, I hope that I can both still use my proxy as well as the github.io domain for projects.